User type implements node
Per
https://gist.github.com/swalkinshaw/3a33e2d292b60e68fcebe12b62bbb3e2#ids-and-the-node-interface,
objects that are persisted ought to have an ID at all times. I'm a bit
hesitant on this because what about cases where we attempt to create an
object, and that object cannot be persisted? Do we merely return null
for the object in those cases?

In any case, this makes the User type a bit more in line with
appropriate graphql api design standards and adjusts the existing things
that return Users to return null in cases the user does not truly exist.

Resolves https://github.com/wecohere/graphql-rails-example/issues/7